Handle photo and video responses. Camera behaviour added
